I did some work on the mistakes

+++ src/com/urise/webapp/storage/SortedArrayStorage.java
@@ -17,14 +17,14 @@ public class SortedArrayStorage extends AbstractArrayStorage {
 
     @Override
     protected void doDelete(Integer index) {
-        System.arraycopy(storage, (Integer) index + 1, storage, (Integer) index, size - (Integer) index - 1);
+        System.arraycopy(storage, index + 1, storage, index, size - index - 1);
         storage[size] = null;
         size--;
     }
 
     @Override
     protected void insertElement(Resume r, Integer index) {
-        int insertIndex = - (Integer) index - 1;
+        int insertIndex = -(Integer) index - 1;
         System.arraycopy(storage, insertIndex, storage, insertIndex + 1, size - insertIndex);
         storage[insertIndex] = r;
     }
